Tidy the code
