+++ compiler/cmd/.gitignore
@@ -1,1 +1,2 @@
 workspace
+dump*
