+++ compiler/cmd/Cargo.toml
@@ -23,6 +23,8 @@ tar = "0.4.38"
 fs_extra = "1.3.0"
 tempfile = "3.5.0"
 clap = { version = "4.2.7", features = ["derive"] }
+log = "0.4.17"
+flexi_logger = "0.25.4"
 
 [build-dependencies]
 anyhow = "1.0.71"
